Include limited history when cloning with commit ID
When cloning with a commit ID, we have to be able to find that commit ID in the
history. Thus, we need history. Therefore, it follows that we can't just have
the super-fast clone depth of 1. However, we probably still don't need the full
depth of commits. Thus, let's download to a depth of 2048 when cloning to a
particular commit ID